ProcessAnaTuple: modified to run with updated AnaTuples
- Moved definition of AnaTupleReader into separate file
- AnaTupleReader: removed branch lists.
- AnaTupleReader: introduced make_function and bind_this to simplify Define calls.
