+++ Analysis/include/EventTags.h
@@ -13,7 +13,7 @@ struct EventTags {
     using DataId = EventAnalyzerDataId;
 
     std::vector<DataId> dataIds;
-    std::vector<double> weights;
+    std::vector<float> weights;
 };
 
 class EventTagCreator {
@@ -26,14 +26,14 @@ public:
                     const std::set<UncertaintySource>& _norm_unc_sources,
                     bool _use_kinFit, bool _use_svFit);
 
-    int CreateVBFTag(const LorentzVectorM& VBF1_p4, const LorentzVectorM& VBF2_p4, bool is_VBF,
-                     bool pass_vbf_trigger) const;
+    static int CreateVBFTag(const LorentzVectorM& VBF1_p4, const LorentzVectorM& VBF2_p4, bool is_VBF,
+                     bool pass_vbf_trigger);
 
-    EventTags CreateEventTags(const std::vector<DataId>& dataIds_base, const std::vector<double>& weights,
+    EventTags CreateEventTags(const std::vector<DataId>& dataIds_base, const std::vector<float>& weights,
                               int num_central_jets, bool has_b_pair, int num_btag_loose, int num_btag_medium,
                               int num_btag_tight, bool is_vbf, bool is_boosted, int vbf_tag_raw,
                               const LorentzVectorM& SVfit_p4, const LorentzVectorM& MET_p4,
-                              double m_bb, double m_tt_vis, int kinFit_convergence) const;
+                              float m_bb, float m_tt_vis, int kinFit_convergence) const;
 
 private:
     const EventCategorySet& categories;
